Refactor feature service to be part of config
