+++ airbyte-webapp/src/config/defaultConfig.ts
@@ -1,5 +1,18 @@
 import { Config } from "./types";
 import { uiConfig } from "./uiConfig";
+import { Feature, FeatureItem } from "hooks/services/Feature";
+
+const Features: Feature[] = [
+  {
+    id: FeatureItem.AllowUploadCustomImage,
+  },
+  {
+    id: FeatureItem.AllowCustomDBT,
+  },
+  {
+    id: FeatureItem.AllowUpdateConnectors,
+  },
+];
 
 const defaultConfig: Config = {
   ui: uiConfig,
@@ -10,6 +23,7 @@ const defaultConfig: Config = {
   integrationUrl: "/docs",
   oauthRedirectUrl: `${window.location.protocol}//${window.location.host}`,
   isDemo: false,
+  features: Features,
 };
 
 export { defaultConfig };
