+++ airbyte-webapp/src/packages/cloud/services/config/index.ts
@@ -1,12 +1,24 @@
 import {
   defaultConfig as coreDefaultConfig,
   useConfig as useCoreConfig,
+  Config,
 } from "config";
 import { CloudConfig, CloudConfigExtension } from "./types";
+import { FeatureItem } from "hooks/services/Feature";
 
 export function useConfig(): CloudConfig {
   return useCoreConfig<CloudConfig>();
 }
+
+const features = [
+  {
+    id: FeatureItem.AllowOAuthConnector,
+  },
+];
+
+const coreDefaultConfigOverrites: Partial<Config> = {
+  features,
+};
 
 const cloudConfigExtensionDefault: CloudConfigExtension = {
   cloudApiUrl: "",
@@ -26,6 +38,7 @@ const cloudConfigExtensionDefault: CloudConfigExtension = {
 export const defaultConfig: CloudConfig = Object.assign(
   {},
   coreDefaultConfig,
+  coreDefaultConfigOverrites,
   cloudConfigExtensionDefault
 );
